fix response

+++ src/controllers/pilots/updatePilotController.js
@@ -4,23 +4,40 @@ import path, { dirname, join } from 'path'
 import { PilotModel } from "../../models/Pilots.js"
 
 const updatePilotController = async (req, res) => {
+    let response = {
+        code: 200,
+        message: "",
+        data: {},
+    }
     try {
         const { id } = req.params
         const { name, phone, nik, certificateExpiredDate } = req.body
-    
-        const certificate = req.files['certificate'][0]
-        const ktpPicture = req.files['ktpPicture'][0]
+
+        const lastPilot = await PilotModel.findById({
+            _id: id,
+            auth: {
+                userId: req.userId,
+                role: req.role
+            }
+        })
+
+        let ktpPicture = lastDrone.document.ktpPicture
+        if (req.files['ktpPicture']) ktpPicture = req.files['ktpPicture'][0]
+
+        let certificate = lastDrone.document.certificate
+        if (req.files['certificate']) certificate = req.files['certificate'][0]
+
         const updatedData = {
             name,
             phone,
             nik,
-            ktpPicture: ktpPicture.path.replace(/\\/g, '/'),
+            ktpPicture: ktpPicture === lastPilot.ktpPicture ? ktpPicture : ktpPicture.path.replace(/\\/g, '/'),
             certificateExpiredDate,
-            certificate: certificate.path.replace(/\\/g, '/'),
+            certificate: certificate === lastPilot.certificate ? certificate : certificate.path.replace(/\\/g, '/'),
             updatedAt: new Date()
         }
 
-        const lastPilot = await PilotModel.findOneAndUpdate(
+        const newPilot = await PilotModel.findOneAndUpdate(
             {
                 _id: id,
                 auth: {
@@ -28,22 +45,37 @@ const updatePilotController = async (req, res) => {
                     role: req.role
                 }
             },
-            updatedData
+            updatedData,
+            { new: true }
         )
+
 
         const __filename = fileURLToPath(import.meta.url)
         const __dirname = dirname(__filename)
         const assetsDirectory = join(__dirname, '../../../')
+        
+        if (ktpPicture !== lastDrone.ktpPicture)
         fs.unlinkSync(join(assetsDirectory, lastPilot.ktpPicture))
+
+        if (certificate !== lastDrone.certificate)
         fs.unlinkSync(join(assetsDirectory, lastPilot.certificate))
 
-        if (!lastPilot) {
-            return res.status(404).json({ error: 'Pilot not found' })
+        if (!newPilot) {
+            response.code = 404
+            response.message = "Pilot not found"
+            response.data = {}
+            return res.status(404).json(response)
         }
 
-        return res.status(200).json({ message: "Pilot updated succesfull" })
-    } catch (error) {
-        res.status(500).json({ error: "Internal server error", detail: error.message })
+        response.code = 200
+        response.message = "Pilot data successfully update"
+        response.data = { pilot: newPilot }
+        return res.status(200).json(response)
+    } catch (e) {
+        response.code = 500
+        response.message = e.message
+        response.data = {}
+        return res.status(500).json(response)
     }
 }
